pyproject.toml
[project]
name = "genaimodule-session3"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "faiss-cpu>=1.12.0",
    "langchain-community>=0.4.1",
    "langchain-openai>=1.0.1",
    "markdown-it-py>=4.0.0",
    "numpy>=2.3.4",
    "openai>=2.6.1",
    "pillow>=11.3.0",
    "pypdf>=6.1.3",
    "python-dotenv>=1.2.1",
    "rank-bm25>=0.2.2",
    "scikit-learn>=1.7.2",
    "sentence-transformers>=5.1.2",
    "streamlit>=1.50.0",
]
